Refactor dataProxy to work with async network calls. Adjust Pages to new async DataProxy. Removed deprecated DownloadManager. Restored progress bar dialog for MAC. Improved the error handling. edit = fixed a bug for building in release mode edit2 = fixed a problem with race conditions in the callbacks of the dataProxy signals edit3 = Removed un-necessary mutex edit4 = Added new JSON parser (pico-json)         fixed some bugs in the genes/datases download         fixed a bug caused but duplicating image textures         fixed a bug with the error text translations         removed network reply auto abort         restored black background color edit5 = Added missed importing of picojson source in CMake edit6 = Setting picojson linking languate to c++ edit7 = Fixed a typo in DataProxy.h and improved the exit 	mechanism in stVi.cpp

+++ src/viewOpenGL/GeneRendererGL.cpp
@@ -292,7 +292,7 @@ void GeneRendererGL::updateVisual()
                 //const qreal adjustment = currentHits / m_max;
                 //In order to adjust by the currentHits, we would need to
                 //know the local max value of the selected genes in this feature...
-                const qreal adjustment = 1 / newRefCount;
+                const qreal adjustment = 1.0 / static_cast<qreal>(newRefCount);
                 dataColor = STMath::lerp(adjustment, dataColor, featureColor);
                 m_geneData.updateQuadColor(index, dataColor);
             }
